+++ htdocs/lib/DemodulatorPanel.js
@@ -50,10 +50,10 @@ function DemodulatorPanel(el) {
 };
 
 DemodulatorPanel.prototype.render = function() {
-    var available = Modes.getModes().filter(function(m){ return m.isAvailable(); });
-    var normalModes = available.filter(function(m){ return m.type === 'analog'; });
-    
-    var digiModes = available
+    var normalModes = Modes.getModes()
+        .filter(function(m){ return m.type === 'analog'; });
+
+    var digiModes = Modes.getModes()
         .filter(function(m){ return m.type === 'digimode'; })
         .sort(function(a, b){ return a.name.localeCompare(b.name) });
 
@@ -95,10 +95,6 @@ DemodulatorPanel.prototype.setMode = function(requestedModulation, underlyingMod
     }
 
     if (this.mode === mode && this.underlyingModulation === underlyingModulation) {
-        return;
-    }
-    if (!mode.isAvailable()) {
-        divlog('Modulation "' + mode.name + '" not supported. Please check the feature report', true);
         return;
     }